Refactors file locations into more sensible places
also renames some of the command line arguments
