--- src/Encrypt.Config.ConsoleHost/ConsoleStateMachine/Factories/WellKnownCommands.cs
@@ -1,10 +1,0 @@
-namespace Encrypt.Config.ConsoleHost.ConsoleStateMachine.Factories {
-    public class WellKnownCommands
-    {
-        public const string CREATE_CONTAINER = "create container";
-        public const string EXPORT_KEY = "export key";
-        public const string IMPORT_KEY = "import key";
-        public const string ENCRYPT_FILE = "encrypt";
-        public const string CREATE_KEYS = "create keys";
-    }
-}
