+++ src/Encrypt.Config.ConsoleHost/Encrypt.Config.ConsoleHost.csproj
@@ -51,12 +51,13 @@
     <Compile Include="ConsoleStateMachine\Factories\ExportStateFactory.cs" />
     <Compile Include="ConsoleStateMachine\Factories\IConsoleStateFactory.cs" />
     <Compile Include="ConsoleStateMachine\Factories\EncryptionStateFactory.cs" />
-    <Compile Include="ConsoleStateMachine\Factories\WellKnownCommands.cs" />
+    <Compile Include="WellKnownCommands.cs" />
     <Compile Include="ConsoleStateMachine\States\ConsoleState.cs" />
     <Compile Include="ConsoleStateMachine\States\CreateKeysState.cs" />
     <Compile Include="ConsoleStateMachine\States\EndState.cs" />
     <Compile Include="ConsoleStateMachine\States\ExportState.cs" />
     <Compile Include="ConsoleStateMachine\States\EncryptionState.cs" />
+    <Compile Include="FileEncrypter.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WellKnownCommandArguments.cs" />
@@ -65,14 +66,6 @@
     <ProjectReference Include="..\Encrypt.Config.Encryption\Encrypt.Config.Encryption.csproj">
       <Project>{32CAF4FC-2216-4964-946B-5E3EA7AAA849}</Project>
       <Name>Encrypt.Config.Encryption</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Encrypt.Config.Json\Encrypt.Config.Json.csproj">
-      <Project>{B5E8B5FA-1E41-47F3-85BA-04FCBCF4ABBA}</Project>
-      <Name>Encrypt.Config.Json</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Encrypt.Config.RSA\Encrypt.Config.RSA.csproj">
-      <Project>{1475C531-0697-42A2-B2F0-1A35A14EAB0B}</Project>
-      <Name>Encrypt.Config.RSA</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
