+++ src/Encrypt.Config.ConsoleHost/WellKnownCommandArguments.cs
@@ -1,15 +1,13 @@
 namespace Encrypt.Config.ConsoleHost {
     public class WellKnownCommandArguments
     {
-        public const string IMPORT_KEY = "i";
-        public const string ENCRYPTED_FILE_OUT = "eo";
-        public const string FILE_PATH = "jc";
-        public const string EXPORT_KEY = "o";
+        public const string IMPORT_KEY = "imp";
+        public const string ENCRYPTED_FILE_OUT = "enc";
+        public const string FILE_PATH = "fp";
+        public const string EXPORT_KEY = "exp";
         public const string EXPORT_PUBLIC_KEY = "pb";
         public const string EXPORT_PRIVATE_KEY = "pv";
-
-        public const string PUBLIC_KEY_PATH = "pi";
-        public const string CONTAINER_NAME = "n";
-        public const string USERNAME = "u";
+        public const string CONTAINER_NAME = "nme";
+        public const string USERNAME = "usr";
     }
 }
