+++ src/Encrypt.Config.ConsoleHost/WellKnownCommands.cs
@@ -1,0 +1,8 @@
+namespace Encrypt.Config.ConsoleHost {
+    public class WellKnownCommands
+    {
+        public const string EXPORT_KEY = "export";
+        public const string ENCRYPT_FILE = "encrypt";
+        public const string DECRYPT_FILE = "decrypt";
+    }
+}
