+++ src/Encrypt.Config.Encryption/NumberGenerators/IGenerateRandomBytes.cs
@@ -1,4 +1,4 @@
-namespace Encrypt.Config.Encryption.Random {
+namespace Encrypt.Config.Encryption.NumberGenerators {
     public interface IGenerateRandomBytes
     {
         byte[] Generate(int length);
