+++ src/Encrypt.Config.Encryption/Symmetric/ISymmetricKeyEncryption.cs
@@ -1,4 +1,4 @@
-namespace Encrypt.Config.Encryption {
+namespace Encrypt.Config.Encryption.Symmetric {
     public interface ISymmetricKeyEncryption
     {
         byte[] Encrypt(byte[] data, byte[] key, byte[] iv);
