+++ tests/Encrypt.Config.Console.Tests/Encryption/HybridEncryptTests/WhenDecryptingData.cs
@@ -1,7 +1,8 @@
 ﻿using System.Security.Cryptography;
 using System.Security.Principal;
+using Encrypt.Config.Encryption;
 using Encrypt.Config.Encryption.Asymmetric;
-using Encrypt.Config.Encryption.RSA;
+using Encrypt.Config.Encryption.Hybrid;
 using Encrypt.Config.Encryption.Symmetric;
 using NUnit.Framework;
