+++ tests/Encrypt.Config.Console.Tests/Encryption/HybridEncryptTests/WhenImportingKey.cs
@@ -1,8 +1,9 @@
 using System.IO;
 using System.Security.Cryptography;
 using System.Security.Principal;
+using Encrypt.Config.Encryption;
 using Encrypt.Config.Encryption.Asymmetric;
-using Encrypt.Config.Encryption.RSA;
+using Encrypt.Config.Encryption.Hybrid;
 using Encrypt.Config.Encryption.Symmetric;
 using NUnit.Framework;
